Move functions into extension
